Refactor touchBegan method to use reference passing rather than duplicating code
The previous iteration of the touchBegan method only checked the side the tap was on before checking
what action to take, duplicating code and making direct references to the GameplayScene's instance
variables. This new version takes advantage of references to avoid code duplication and instead uses
variables that exist only within the scope of the method to determine what action to take.